Add more unit tests and increate code reading. The responsibility of file-reader.js is only read files. A separeted function was created to get the content of the license. Connects to #211
