+++ lib/file-reader.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const fs = require('graceful-fs');
-var request = require('sync-request');
+const remoteLicense = require('./remote-license');
 
 /**
  * This function reads a file and return null in case
@@ -16,13 +16,7 @@ function readListFile (file) {
 
 function readAsJson (url) {
   if (url.startsWith('http')) {
-    const res = request('GET', url);
-    if (res.statusCode === 200) {
-      return JSON.parse(res.body.toString());
-    } else {
-      console.log('Could not get resource: ', url);
-      console.log('StatusCode returned was: ', res.statusCode);
-    }
+    return remoteLicense.fetch(url);
   } else {
     const content = readListFile(url);
     if (content !== null) {
