+++ package.json
@@ -4,7 +4,7 @@
   "description": "License-reporter gathers license information and reports them in various formats.",
   "main": "index.js",
   "scripts": {
-    "test": "nyc tape test/*.js | tap-spec",
+    "test": "nyc tape test/*.js | tap-diff",
     "coverage": "nyc report --reporter=text-lcov > ./coverage/lcov.info",
     "lint": "eslint test/*.js index.js lib/*.js bin/*.js",
     "docs": "./node_modules/.bin/jsdoc --verbose -d docs -t ./node_modules/ink-docstrap/template -R README.md index.js ./lib/*.js",
@@ -50,6 +50,7 @@
     "proxyquire": "^1.8.0",
     "rewire": "^2.5.2",
     "standard-version": "^3.0.0",
+    "tap-diff": "^0.1.1",
     "tap-spec": "^4.1.1",
     "tape": "^4.8.0",
     "test-console": "^1.0.0"
