+++ test/file-reader-test.js
@@ -2,7 +2,6 @@
 
 const test = require('tape');
 const path = require('path');
-const proxyquire = require('proxyquire');
 const reader = require('../lib/file-reader.js');
 
 test('Should return null when file not found.', (t) => {
@@ -42,29 +41,5 @@ test('Should return file as JSON.', (t) => {
 test('Should return null file not found (JSON).', (t) => {
   t.plan(1);
   t.equal(null, reader.readAsJson('something'));
-  t.end();
-});
-
-test('Should return null if http resource not found.', (t) => {
-  const requestStub = function (method, url) {
-    return { statusCode: 404 };
-  };
-  const reader = proxyquire('../lib/file-reader.js', { 'sync-request': requestStub });
-  t.plan(1);
-  t.equal(null, reader.readAsJson('http://bogus.com/namemap.json'));
-  t.end();
-});
-
-test('Should return JSON if http resource is found.', (t) => {
-  const requestStub = function (method, url) {
-    return {
-      statusCode: 200,
-      body: `{ "name": "something"}`
-    };
-  };
-  const reader = proxyquire('../lib/file-reader.js', { 'sync-request': requestStub });
-  t.plan(1);
-  const json = reader.readAsJson('http://bogus.com/namemap.json');
-  t.equal(json.name, 'something');
   t.end();
 });
